Avoid returning duplicate definitions
When all definitions of a specific type are retrieved from a
FHIRDefinitions object, the returned array will not contain duplicates.
The reported size of the FHIRDefinitions object will count each
contained definition only once, regardless of the number of times it
appears.

The DoubleMap facilitates maintaining a list of unique definitions. If
child definitions exist, though, it is necessary to check the resulting
list for uniqueness.