Dependency version can be "latest" (#13)
* Dependency version can be "latest"

When the version is "latest", try to get the latest version. Then, use
that version when loading the package. If the latest version is
unavailable, an error is thrown. The version will be checked against a
custom package registry if one has been specified.

* Add getLatestVersion to api exports

Function renamed in load.ts to avoid name collision.

* Consistent error handling in lookUpLatestVersion

If the axios requests to get package information fail, throw
LatestVersionUnavailableError instead of rethrowing the error from
axios. LatestVersionUnavailableError is now the only error type that
should be thrown from this function.

* Add getLatestVersion to README