Downgrade download error to warning if package can be loaded from cache
In the case of #current packages, FPL always attempts to lookup and download the latest build (if stale) -- even if there is a cached copy in the FHIR cache. Previously, if this lookup/download failed, we logged an error; now, we log a warning. This change was discussed on Zulip here: https://chat.fhir.org/#narrow/channel/179252-IG-creation/topic/Issue.20using.20version-specific.20packages/near/513733847

Fixes #55